refactor(installpackage): make install package more readable
