+++ packages/core/src/package/packageInstallers/InstallPackage.ts
@@ -2,7 +2,7 @@ import SFPLogger, { COLOR_KEY_MESSAGE, Logger, LoggerLevel } from '@dxatscale/sf
 import { PackageInstallationResult, PackageInstallationStatus } from './PackageInstallationResult';
 import ProjectConfig from '../../project/ProjectConfig';
 import SFPStatsSender from '../../stats/SFPStatsSender';
-import PackageInstallationHelpers from './PackageInstallationHelpers';
+import AssignPermissionSets from '../../permsets/AssignPermissionSets';
 import ScriptExecutor from '../../scriptExecutor/ScriptExecutorHelpers';
 import { Connection } from '@salesforce/core';
 import * as fs from 'fs-extra';
@@ -40,7 +40,6 @@ export class SfpPackageInstallationOptions {
 }
 
 export abstract class InstallPackage {
-
     protected connection: Connection;
     protected packageDescriptor;
     protected packageDirectory;
@@ -56,8 +55,8 @@ export abstract class InstallPackage {
 
     public async exec(): Promise<PackageInstallationResult> {
         let startTime = Date.now();
-        let elapsedTime:number;
-        try {        
+        let elapsedTime: number;
+        try {
             this.packageDescriptor = ProjectConfig.getSFDXPackageDescriptor(
                 this.sfpPackage.sourceDir,
                 this.sfpPackage.packageName
@@ -67,46 +66,55 @@ export abstract class InstallPackage {
 
             if (await this.isPackageToBeInstalled(this.options.skipIfPackageInstalled)) {
                 if (!this.options.isDryRun) {
-                    //Package Has Permission Set Group
                     await this.waitTillAllPermissionSetGroupIsUpdated();
                     await this.assignPermsetsPreDeployment();
                     await this.executePreDeploymentScripts();
-                    await this.getPackageDirectoryForAliasifiedPackages();
+                    await this.setPackageDirectoryForAliasifiedPackages();
                     await this.install();
                     await this.assignPermsetsPostDeployment();
                     await this.executePostDeployers();
                     await this.executePostDeploymentScript();
                     await this.commitPackageInstallationStatus();
 
-
-                    elapsedTime=Date.now()-startTime;
+                    elapsedTime = Date.now() - startTime;
                     this.sendMetricsWhenSuccessfullyInstalled(elapsedTime);
                 }
-                return { result: PackageInstallationStatus.Succeeded , elapsedTime:elapsedTime};
+                return { result: PackageInstallationStatus.Succeeded, elapsedTime: elapsedTime };
             } else {
                 SFPLogger.log('Skipping Package Installation', LoggerLevel.INFO, this.logger);
                 return { result: PackageInstallationStatus.Skipped };
             }
         } catch (error) {
-            elapsedTime=Date.now()-startTime;
+            elapsedTime = Date.now() - startTime;
             this.sendMetricsWhenFailed(elapsedTime);
             return {
                 result: PackageInstallationStatus.Failed,
                 message: error.message,
-                elapsedTime:elapsedTime
+                elapsedTime: elapsedTime,
             };
         }
     }
 
     private async waitTillAllPermissionSetGroupIsUpdated() {
-        let permissionSetGroupUpdateAwaiter: PermissionSetGroupUpdateAwaiter = new PermissionSetGroupUpdateAwaiter(
-            this.connection,
-            this.logger
-        );
-        await permissionSetGroupUpdateAwaiter.waitTillAllPermissionSetGroupIsUpdated();
+        try {
+            //Package Has Permission Set Group
+            let permissionSetGroupUpdateAwaiter: PermissionSetGroupUpdateAwaiter = new PermissionSetGroupUpdateAwaiter(
+                this.connection,
+                this.logger
+            );
+            await permissionSetGroupUpdateAwaiter.waitTillAllPermissionSetGroupIsUpdated();
+        } catch (error) {
+            //Ignore error
+            // Lets try proceeding
+            SFPLogger.log(
+                `Unable to check the status of Permission Set Groups due to ${error}`,
+                LoggerLevel.WARN,
+                this.logger
+            );
+        }
     }
 
-    protected async getPackageDirectoryForAliasifiedPackages() {
+    protected async setPackageDirectoryForAliasifiedPackages() {
         if (this.packageDescriptor.aliasfy) {
             const searchDirectory = path.join(this.sfpPackage.sourceDir, this.packageDescriptor.path);
             const files = FileSystem.readdirRecursive(searchDirectory, true);
@@ -150,7 +158,7 @@ export abstract class InstallPackage {
         }
     }
 
-    private sendMetricsWhenFailed(elapsedTime:number) {
+    private sendMetricsWhenFailed(elapsedTime: number) {
         SFPLogger.log(
             `Package ${COLOR_KEY_MESSAGE(
                 this.sfpPackage.package_name
@@ -163,7 +171,7 @@ export abstract class InstallPackage {
         });
     }
 
-    private sendMetricsWhenSuccessfullyInstalled(elapsedTime:number) {
+    private sendMetricsWhenSuccessfullyInstalled(elapsedTime: number) {
         SFPLogger.log(
             `Package ${COLOR_KEY_MESSAGE(this.sfpPackage.package_name)} installation took ${COLOR_KEY_MESSAGE(
                 getFormattedTime(elapsedTime)
@@ -209,18 +217,21 @@ export abstract class InstallPackage {
         } else return true; // Always install packages if skipIfPackageInstalled is false
     }
 
+    private async assignPermsetsPreDeployment() {
+        try {
+            if (this.sfpPackage.assignPermSetsPreDeployment) {
+                SFPLogger.log('Assigning permission sets before deployment:', LoggerLevel.INFO, this.logger);
 
-    private async assignPermsetsPreDeployment()
-    {
-        if (this.sfpPackage.assignPermSetsPreDeployment) {
-            SFPLogger.log('Assigning permission sets before deployment:', LoggerLevel.INFO, this.logger);
-
-            await PackageInstallationHelpers.applyPermsets(
-                this.sfpPackage.assignPermSetsPreDeployment,
-                this.connection,
-                this.sfpPackage.sourceDir,
-                this.logger
-            );
+                await AssignPermissionSets.applyPermsets(
+                    this.sfpPackage.assignPermSetsPreDeployment,
+                    this.connection,
+                    this.sfpPackage.sourceDir,
+                    this.logger
+                );
+            }
+        } catch (error) {
+            //Proceed ahead not a critical error to break installation
+            SFPLogger.log(`Unable to assign permsets (Pre Deployment) due to ${error}`, LoggerLevel.WARN, this.logger);
         }
     }
 
@@ -243,17 +254,21 @@ export abstract class InstallPackage {
 
     abstract install();
 
-    private async assignPermsetsPostDeployment()
-    {
-        if (this.sfpPackage.assignPermSetsPostDeployment) {
-            SFPLogger.log('Assigning permission sets after deployment:', LoggerLevel.INFO, this.logger);
+    private async assignPermsetsPostDeployment() {
+        try {
+            if (this.sfpPackage.assignPermSetsPostDeployment) {
+                SFPLogger.log('Assigning permission sets after deployment:', LoggerLevel.INFO, this.logger);
 
-            await PackageInstallationHelpers.applyPermsets(
-                this.sfpPackage.assignPermSetsPostDeployment,
-                this.connection,
-                this.sfpPackage.sourceDir,
-                this.logger
-            );
+                await AssignPermissionSets.applyPermsets(
+                    this.sfpPackage.assignPermSetsPostDeployment,
+                    this.connection,
+                    this.sfpPackage.sourceDir,
+                    this.logger
+                );
+            }
+        } catch (error) {
+            //Proceed ahead not a critical error to break installation
+            SFPLogger.log(`Unable to assign permsets (Post Deployment) due to ${error}`, LoggerLevel.WARN, this.logger);
         }
     }
